requirements.txt
flask==2.0.1
flask-cors==3.0.10
streamlit==1.22.0
google-generativeai>=0.3.0
python-dotenv==0.19.0
requests==2.31.0
pytest==7.3.1
pyperclip==1.8.2